Extract ESDocument to a separate file and add safeguard in __init__
