+++ connectors/es/__init__.py
@@ -4,5 +4,6 @@
 # you may not use this file except in compliance with the Elastic License 2.0.
 #
 from connectors.es.client import ESClient  # NOQA
+from connectors.es.document import ESDocument, InvalidDocumentSourceError  # NOQA
 from connectors.es.index import ESIndex  # NOQA
 from connectors.es.settings import DEFAULT_LANGUAGE, Mappings  # NOQA
